fix: move common definitions to common.rs
