+++ fix/runtime/src/lib.rs
@@ -7,3 +7,4 @@ pub mod bottom;
 pub mod fixruntime;
 pub mod runtime;
 pub mod storage;
+pub mod common;
